% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lp.R
\name{add_constraints}
\alias{add_constraints}
\title{Add constraints to a tidy_lp}
\usage{
add_constraints(tlp, ...)
}
\arguments{
\item{tlp}{a tidy_lp object}

\item{...}{unnamed constraints specified as formulas}
}
\description{
Add constraints to a tidy_lp
}
