% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lp.R
\name{bind_solution}
\alias{bind_solution}
\title{Bind the solution to the original data}
\usage{
bind_solution(tidy_lp_solution, filter_nonzero = FALSE, name = ".solution")
}
\arguments{
\item{tidy_lp_solution}{a solved TidyLPSolution}

\item{filter_nonzero}{if true, remove rows corresponding to 0's in the solution}

\item{name}{the name of the solution column}
}
\description{
Returns the original supplied data frame with a new column containing the solution.
}
