% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lp.R
\name{categorical_constraint}
\alias{categorical_constraint}
\title{Categorical Constraints}
\usage{
categorical_constraint(col)
}
\arguments{
\item{col}{the column to generate a categorical constraint set from}
}
\value{
a tibble
}
\description{
Categorical Constraints
}
