% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lp.R
\docType{data}
\name{fantasy_points}
\alias{fantasy_points}
\title{Fantasy Points}
\format{
\subsection{\code{fantasy_points}}{

A data frame with 322 rows and 6 columns.

\describe{
\item{Id}{A unique identifier for the player}
\item{Position}{The player's position}
\item{Name}{The player's name}
\item{Projected_FP}{The player's projected fantasy points}
\item{Salary}{The player's salary}
\item{Team}{The player's NBA team}
}
}
}
\usage{
fantasy_points
}
\description{
A modified and stripped down version of a fantasy basketball dataset originating from Fanduel.
}
\keyword{datasets}
