% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lp.R
\name{leq}
\alias{leq}
\alias{geq}
\alias{eq}
\title{Constraint direction helpers}
\usage{
leq(value)

geq(value)

eq(value)
}
\arguments{
\item{value}{the right hand side of the constraint}
}
\value{
a \code{constraint_rhs} object
}
\description{
These functions are used to specify the direction of the constraint
formulas. These should only be used within a constraint formula.

\code{geq} specifies a >= constraint
\code{leq} specifies a <= constraint
\code{eq} denotes a == constraint
}
