% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lp.R
\name{lp_solve}
\alias{lp_solve}
\title{Solve a TidyLP}
\usage{
lp_solve(tlp, ..., solver = "lpSolve")
}
\arguments{
\item{tlp}{a TidyLP object}

\item{...}{additional arguments passed to the solver}

\item{solver}{the solver to use. Currently does nothing as only lpSolve is implemented}
}
\value{
the solution
}
\description{
Solve a TidyLP
}
