% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lp.R
\name{pretty_constraint}
\alias{pretty_constraint}
\title{String representation of a constraint}
\usage{
pretty_constraint(x)
}
\arguments{
\item{x}{a constraint}
}
\value{
a string
}
\description{
String representation of a constraint
}
