% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lp.R
\name{pretty_tlp}
\alias{pretty_tlp}
\title{A pretty representation of a tidy LP problem}
\usage{
pretty_tlp(x)
}
\arguments{
\item{x}{a tidy LP problem}
}
\value{
as string
}
\description{
A pretty representation of a tidy LP problem
}
