% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lp.R
\name{tidy_lp}
\alias{tidy_lp}
\title{Title}
\usage{
tidy_lp(
  .data,
  objective_coefs,
  const_lhs,
  const_rhs,
  direction = c("max", "min"),
  const_dir = NULL,
  obj_labels = NULL
)
}
\arguments{
\item{direction}{}
}
\value{

}
\description{
Title
}
