% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyLP-package.R
\docType{package}
\name{tidyLP-package}
\alias{tidyLP}
\alias{tidyLP-package}
\title{tidyLP: Tidy Linear Programs}
\description{
A tidy interface for linear programming.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/colin-fraser/tidyLP}
  \item Report bugs at \url{https://github.com/colin-fraser/tidyLP/issues}
}

}
\author{
\strong{Maintainer}: Colin Fraser \email{colin.r.fraser@gmail.com}

}
\keyword{internal}
