% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lp.R
\name{tidy_lp}
\alias{tidy_lp}
\title{Create a TidyLP object}
\usage{
tidy_lp(
  .data,
  .objective,
  ...,
  .direction = "max",
  .all_int = FALSE,
  .all_bin = FALSE
)
}
\arguments{
\item{.data}{a data.frame that will be used to construct the object}

\item{.objective}{an expression for the objective function}

\item{...}{constraints}

\item{.direction}{the direction of obtimization}

\item{.all_int}{should the solution be integer-valued?}

\item{.all_bin}{should the solution be binary-valued?}
}
\value{
a TidyLPProblem object
}
\description{
Create a TidyLP object
}
